% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{chart}
\usage{
chart(input, output, session, g, setdrill = NULL, noopt = 0)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is the graph/chart to be charted}

\item{setdrill}{is the function chart will call upstream to set a drill value on a chart}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
A 'shiny' module to display many types of charts available as 'htmlwidgets' with a dataframe as input
}
\details{
The graph structure containing chart type and chart data is passed as input.

The graph/chart data is displayed in appropriate chart type with options to change to other chart types
}
\seealso{
See chartex for an example
}
