% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\docType{data}
\name{cr}
\alias{cr}
\title{cr}
\format{
An object of class \code{reactivevalues} of length 3.
}
\usage{
cr
}
\description{
cr is chart reactive that tracks user click on chart types
}
\keyword{datasets}
