% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ctreeUI}
\alias{ctreeUI}
\title{ctreeUI}
\usage{
ctreeUI(id, g, noopt = 0)
}
\arguments{
\item{id}{is the caller's id}

\item{g}{is the graph/chart to be charted}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
User interface to display 'collapsibleTree' chart type
}
