% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl2g.R
\name{df2g}
\alias{df2g}
\title{df2g}
\usage{
df2g(title, dxy, ndim = 1, nseries = 1)
}
\arguments{
\item{title}{is the title for the chart}

\item{dxy}{is the dataframe to draw the chart from}

\item{ndim}{is the number of dimensions in the xl file; it is assumed these are in the first ndim columns of the xl}

\item{nseries}{is the number of series in the xl file; it is assumed these are in the last nseries columns of the xl}
}
\description{
This is a utility function to create a 'graph' data structure to pass to chart module
}
