% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyg.R
\name{dyg}
\alias{dyg}
\title{dyg}
\usage{
dyg(input, output, session, g, noopt = 0)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is the graph/chart to be charted}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
A 'shiny' module to display 'dygraph' chart with options
}
\details{
Options for 'dygraph' are range selector and line fill
}
