% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly}
\alias{plotly}
\title{plotly}
\usage{
plotly(input, output, session, g, setdrill = NULL, noopt = 0)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is the graph/chart to be charted}

\item{setdrill}{is the function to chart will call upstream to set a drill value on a chart}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
A 'shiny' module to display 'plot_ly' chart with options
}
\details{
Options for 'plotly' are provided for bar, line, scatter and pie charts

Common options are tick angle for x axis, margin width/height, vertical or horizontal orientation, color/palette

Additional option for bar chart option is stacked bar chart

Additional options for line chart option are line type, line shape, and area fill

Additional options for pie chart option are donut size and clockwise/counter-clockwise drawing
}
