% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunb.R
\name{sunb}
\alias{sunb}
\title{sunb}
\usage{
sunb(input, output, session, g, noopt = 0)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is the graph/chart to be charted}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
A 'shiny' module to display 'sunburst' chart with options
}
\details{
Option for 'sunburst' is color palette
}
