% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{tree}
\usage{
tree(input, output, session, g, noopt = 0)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is the graph/chart to be charted}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
A 'shiny' module to display 'treemap' chart with options
}
\details{
Options for treemap are: interactive or static tree, and choice of color palette
}
