% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbox.R
\name{vbox}
\alias{vbox}
\title{vbox}
\usage{
vbox(input, output, session, g, noopt = 0)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{g}{is the graph/chart to be charted}

\item{noopt}{is a toggle that tells chart module not to display options to change chart defaults}
}
\description{
A 'shiny' module to display 'valueBox' chart with options
}
\details{
This is drawn as one standard value box, with no further options
}
