% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl2g.R
\name{xl2g}
\alias{xl2g}
\title{xl2g}
\usage{
xl2g(xl, ndim = 1, nseries = 1)
}
\arguments{
\item{xl}{has the title and data of the 'Excel' file}

\item{ndim}{is the number of dimensions in the 'Excel' file; it is assumed these are in the first ndim columns of the xl}

\item{nseries}{is the number of series in the 'Excel' file; it is assumed these are in the last nseries columns of the xl}
}
\description{
This is a utility function to create a 'graph' data structure to pass to chart module from an 'Excel' sheet
}
